refactor(component): reflect feedbacks(refactor forms, etc)
- useForm으로 config/phone-number의 Form 로직을 재사용했다.
- LoginForm은 이미 충분히 구체적인 템플릿이므로 전체 페이지 렌더링이 일어나지 않게 로직을 모두 넣어주었다.
- 기타 컨벤션에 따른 코드를 수정했다.
