--- apps/web/pages/login/kakao/index.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function KakaoLoginPage() {
-  return <div>KakaoLoginPage</div>;
-}
-
-export default KakaoLoginPage;
