--- apps/web/pages/login/naver/index.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function NaverLoginPage() {
-  return <div>NaverLoginPage</div>;
-}
-
-export default NaverLoginPage;
