+++ packages/ui/text/index.ts
@@ -2,3 +2,4 @@ export * from './Strong';
 export * from './ScreenReader';
 export * from './Default';
 export * from './IconHeaderText';
+export * from './types';
