+++ packages/ui/text/types.ts
@@ -1,0 +1,1 @@
+export type DynamicHeaderTextType = 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6';
