--- packages/ui/toast/Info.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function Info() {
-  return <div>Info</div>;
-}
